fix(firestore-bigquery-export): Remove broken partition table check
This commit removes a misplaced guard found in the isTablePartitioned method of the Partitioning class. The removed conditional was erroneously checking the truthfulness of a result of a method which returns a promise, thereby always evaluating to being true. Additionally, the removed logging operation is unrelated to the purpose of the method. This errant code likely ended up here as a result of a failed merge. This addresses an issue where firestore-bigquery-export needlessly fills the logs with errors about not being able to partition an existing table even when partitioning is disabled.

The log message called upon by the removed guard is also removed as it is no longer used. It may want to be reintroduced as part of the partitioning table creation phase in a future update, but that is outside of the scope of this update.

Additional discussion found in #1368